requirements.txt
alembic==1.15.1
blinker==1.9.0
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cryptography==44.0.2
cycler==0.12.1
Flask==3.0.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.56.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
lxml==5.3.1
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.8.2
numpy==1.26.4
packaging==24.2
pandas==2.2.1
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==11.1.0
pycparser==2.22
pyparsing==3.2.1
pypdfium2==4.30.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
SQLAlchemy==2.0.27
supabase==2.3.5
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.0.1
WTForms==3.2.1
zipp==3.21.0
reportlab==4.1.0
